backend/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# Data processing
numpy==1.25.2
pandas==2.1.1
scikit-learn==1.3.1
tensorflow==2.14.0
torch==2.1.0
statsmodels==0.14.0

# Stream processing
apache-flink==1.17.1
pyspark==3.4.1
kafka-python==2.0.2

# Time series databases
influxdb-client==1.36.1
elasticsearch==8.10.0
psycopg2-binary==2.9.9  # For TimescaleDB

# Observability
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0

# Utilities
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
